+++ source/PhysicalQuantity/DimensionlessCoefficient.php
@@ -15,12 +15,10 @@ class DimensionlessCoefficient extends AbstractBasePhysicalQuantity
 
     protected static $hasBeenInitialized = false;
 
-    protected static $nativeUnitOfMeasure;
-
     protected static function initializeUnitsOfMeasure()
     {
         $coefficient = UnitOfMeasure::nativeUnitFactory('');
-        static::registerNativeUnitOfMeasure($coefficient);
+        static::registerUnitOfMeasure($coefficient);
     }
 
     public function __construct($value)