+++ tests/Fixtures/PhysicalQuantity/Woogosity.php
@@ -9,14 +9,12 @@ class Woogosity extends AbstractBasePhysicalQuantity
 {
     protected static $unitDefinitions    = [];
     protected static $hasBeenInitialized = false;
-    protected static $nativeUnitOfMeasure;
-
     protected static function initializeUnitsOfMeasure()
     {
         $native = UnitOfMeasure::nativeUnitFactory('l');
         $native->addAlias('lupee');
         $native->addAlias('lupees');
-        static::registerNativeUnitOfMeasure($native);
+        static::registerUnitOfMeasure($native);
 
         $unit = UnitOfMeasure::linearUnitFactory('p', 1.234);
         $unit->addAlias('plurp');